[cleanup] Misc fixes
Closes #4027
